Fix circular import: home shared helper under autolens.lens.plot
The previous commit placed _compute_critical_curve_lines in
autolens/plot/plot_utils.py. Importing any autolens.plot.* submodule runs the
autolens/plot/__init__.py aggregator, which imports the fit-plot modules — so
having a fit-plot module import from autolens.plot.plot_utils created a circular
import that broke 'import autolens' entirely (subplot_fit imported from a
partially-initialised module).

Relocate the canonical definition to autolens/lens/plot/critical_curves.py.
The autolens.lens.plot and autolens.lens package __init__ files are empty, so
importing it from the fit-plot modules triggers no aggregator and no cycle.
Critical curves are a lensing concept, so this is also the more natural home.
plot_utils.py is reverted to its original one-line re-export. tracer_plots.py's
lazy import is repointed at the canonical home, and the witness test patches/
observes autolens.lens.plot.critical_curves.

Verified: 'import autolens' (incl. the autolens.plot aggregator) succeeds, and
the imaging/interferometer fit-plot, tracer-plot, and imaging/interferometer/
point model test suites all pass.
